refactor: move strictly typed code to its own folder
#32
