+++ src/index.ts
@@ -1,9 +1,6 @@
 export * from "./criteria";
 export * from "./property";
-export * from "./query";
+export * from "./strict";
 export * from "./selection";
 export * from "./utils";
 
-export * from "./entity-criteria";
-export * from "./instance";
-export * from "./workspace";
