+++ src/strict/hydratable-query-result.ts
