+++ src/strict/index.ts
@@ -1,4 +1,6 @@
+export * from "./entity-criteria";
 export * from "./hydratable-query-result";
+export * from "./instance";
 export * from "./payload-hydration";
 export * from "./payload-hydrator";
 export * from "./query";
@@ -6,3 +8,4 @@ export * from "./query-result";
 export * from "./query-stream";
 export * from "./query-stream-packet";
 export * from "./query-translator";
+export * from "./workspace";
