+++ src/strict/instance.ts
@@ -1,5 +1,5 @@
-import { Primitive, Unbox } from "./utils";
-import { Property, Attribute, Context } from "./property";
+import { Primitive, Unbox } from "../utils";
+import { Property, Attribute, Context } from "../property";
 
 type BoxPropertyValue<P, V> = P extends Attribute.IsIterable ? V[] : V;
