+++ src/strict/payload-hydrator.ts
