+++ src/strict/query-result.ts
