+++ src/strict/query-stream-packet.ts
