+++ src/strict/query-stream.ts
