+++ src/strict/query-translator.ts
